// server/types/developer.types.ts
type Developer = {
	id?: string
	date?: string
	status: string | Boolean
	sales_name: string
	developer_name: string
	developer_img: string
	phone: string | number
	business: string
	com_company: string
	com_sales: string
	email: string
	password: string
	location_dev: string[]
}
export default Developer